catalog/pyproject.toml
[project]
name = "catalog"
version = "0.1.0"
requires-python = "==3.13.5"
dependencies = [
    "loguru==0.7.3",
    "colorama==0.4.6",
    "fastapi==0.116.0",
    "uvicorn==0.35.0",
    "chardet==5.2.0",
    "Authlib==0.14.3",
    "itsdangerous==1.1.0",
    "passlib==1.7.2",
    "pydantic==2.11.7",
    "bcrypt==4.3.0",
    "valkey==6.1.0",
    "redis==6.4.0",
    "SQLAlchemy==2.0.41",
    "psycopg[binary]==3.2.9",
    "asyncpg==0.30.0",
    "python-jose[cryptography]==3.3.0",
    "cryptography==42.0.8",
    "dotenv==0.9.9",
    "aiobotocore==2.24.1",
    "pillow==11.3.0",
    "python-multipart==0.0.20",
    "bs_models @ git+https://github.com/dagahan/bs_models.git",
    "mypy>=1.17.1",
]


[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "B", "UP", "I", "W", "N", "ANN", "TCH"]
ignore = ["ANN401", "W293", "F403", "F405"]
unfixable = ["ANN"]


[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
show_error_codes = true
pretty = true

disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
no_implicit_optional = true
warn_return_any = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true

plugins = [
  "pydantic.mypy",
  "sqlalchemy.ext.mypy.plugin", 
]

[mypy-aiobotocore]
ignore_missing_imports = true
[mypy-botocore]
ignore_missing_imports = true
[mypy-loguru]
ignore_missing_imports = true
[mypy-starlette]
ignore_missing_imports = true
[mypy-fastapi]
ignore_missing_imports = true


[jwt]
private_key = "certs/jwt-private.pem"


[db]
echo = false
pool_size = 5
max_overflow = 10


